requirements.txt
affine==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
aiosqlite==0.21.0
alphashape==1.3.1
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
asyncio==3.4.3
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.14.2
bleach==6.2.0
blis==1.3.3
Brotli==1.1.0
catalogue==2.0.10
cdsapi==0.7.7
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
chonkie==1.1.2
choreographer==1.2.1
click==8.2.1
click-log==0.4.0
click-plugins==1.1.1.2
cligj==0.7.2
cloudpathlib==0.23.0
cloudscraper==1.2.71
colorama==0.4.6
comm==0.2.3
confection==0.1.5
contourpy==1.3.2
Crawl4AI==0.8.6
cryptography==46.0.3
cssselect==1.3.0
curl_cffi==0.13.0
cycler==0.12.1
cymem==2.0.13
datasets==3.6.0
DateTime==5.5
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dotenv==0.9.9
ecmwf-datastores-client==0.4.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.2.0
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.2
fastuuid==0.14.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fonttools==4.58.2
fqdn==1.5.1
frozendict==2.4.7
frozenlist==1.6.2
fsspec==2025.3.0
gql==3.5.3
graphql-core==3.2.6
greenlet==3.2.3
h11==0.16.0
h2==4.3.0
hf-xet==1.1.3
hnswlib==0.8.0
holidays==0.86
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.26.5
humanize==4.12.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jcs==0.2.1
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaleido==1.2.0
kiwisolver==1.4.8
langdetect==1.0.9
langid==1.1.6
lark==1.3.1
litellm==1.72.2
llama_cpp_python==0.3.16
logistro==2.0.1
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.3
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
multidict==6.4.4
multiprocess==0.70.16
multitasking==0.0.12
multiurl==0.3.7
murmurhash==1.0.15
narwhals==2.1.2
nats-py==2.14.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
newspaper3k==0.2.8
nltk==3.9.1
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.4.1
-e git+ssh://git@github.com/vsevolodnedora/nlp_news_summary.git@a31f4bfe2db0fac55e0c1510a5a1bad1dedf56cd#egg=open_event_intel
openai==2.30.0
openpyxl==3.1.5
orjson==3.11.5
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.4
patchright==1.57.2
pdfminer.six==20251230
pdfplumber==0.11.9
peewee==4.0.3
pendulum==3.1.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.8
playwright==1.52.0
playwright-stealth==2.0.2
plotly==6.3.0
pluggy==1.6.0
preshed==3.0.12
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==7.34.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.1
pyOpenSSL==25.3.0
pyparsing==3.2.3
pypdf==6.9.2
pypdfium2==5.6.0
pyperclip==1.9.0
PySocks==1.7.1
pytest==8.4.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
rank-bm25==0.2.2
RapidFuzz==3.13.0
rasterio==1.4.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.1
rtree==1.4.1
ruff==0.12.5
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
selenium==4.38.0
Send2Trash==1.8.3
sgmllib3k==1.0.0
shapely==2.1.2
simplejson==3.20.2
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.8
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlight==2.0.1
srsly==2.5.2
stack-data==0.6.3
starlette==0.46.2
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
tf-playwright-stealth==1.1.2
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.11.0
tinycss2==1.4.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.11.0
trio==0.32.0
trio-websocket==0.12.2
typer-slim==0.20.1
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
unclecode-litellm==1.81.13
uri-template==1.3.0
urllib3==2.5.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.3
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wrapt==2.0.1
wsproto==1.3.2
xxhash==3.5.0
yarl==1.20.0
yfinance==1.2.0
zipp==3.22.0
zope.interface==7.2
